+++ public-stackql-deploy-install.sh
@@ -1,0 +1,7 @@
+rm -rf stackql-deploy
+rm -rf stackql
+rm -rf stackql-aws-cloud-shell.sh
+rm -rf stackql-azure-cloud-shell.sh
+rm -rf stackql-google-cloud-shell.sh
+rm -rf stackql-databricks-shell.sh
+curl -L https://get-stackql-deploy.io | tar xzf -